requirements.txt
# Core
numpy==2.2.6
tqdm==4.67.3
PyYAML==6.0.3

# Audio / signal
librosa==0.11.0
soundfile==0.13.1
pydub==0.25.1
jiwer==4.0.0

# Model runtime
torch==2.8.0
torchaudio==2.8.0
transformers==4.57.3
accelerate==1.12.0
tokenizers==0.22.2
sentencepiece==0.2.1
safetensors==0.7.0
huggingface_hub==0.36.2

# API backends
openai==2.17.0
google-generativeai==0.8.6

# Translation metrics
sacrebleu==2.6.0
nltk==3.9.2
bert-score==0.3.13
scikit-learn==1.7.2
scipy==1.15.3

# Optional (only for vLLM backends)
# vllm==0.16.0
# vllm-omni==0.16.0
# qwen-omni-utils==0.0.8

